+++ src/main/java/org/websoso/WSSServer/application/LibraryEvaluationApplication.java
@@ -3,8 +3,12 @@ package org.websoso.WSSServer.application;
 import static org.websoso.WSSServer.exception.error.CustomUserNovelError.USER_NOVEL_ALREADY_EXISTS;
 
 import jakarta.transaction.Transactional;
+import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
+import java.util.Map;
+import java.util.Set;
+import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
@@ -12,14 +16,22 @@ import org.websoso.WSSServer.domain.User;
 import org.websoso.WSSServer.dto.keyword.KeywordGetResponse;
 import org.websoso.WSSServer.dto.userNovel.UserNovelCreateRequest;
 import org.websoso.WSSServer.dto.userNovel.UserNovelGetResponse;
+import org.websoso.WSSServer.dto.userNovel.UserNovelUpdateRequest;
 import org.websoso.WSSServer.exception.exception.CustomUserNovelException;
+import org.websoso.WSSServer.library.domain.AttractivePoint;
+import org.websoso.WSSServer.library.domain.Keyword;
 import org.websoso.WSSServer.library.domain.UserNovel;
+import org.websoso.WSSServer.library.domain.UserNovelAttractivePoint;
+import org.websoso.WSSServer.library.domain.UserNovelKeyword;
 import org.websoso.WSSServer.library.service.AttractivePointService;
 import org.websoso.WSSServer.library.service.KeywordService;
 import org.websoso.WSSServer.library.service.LibraryService;
 import org.websoso.WSSServer.novel.domain.Novel;
 import org.websoso.WSSServer.novel.service.NovelServiceImpl;
 
+/**
+ * 서재 평가는 서재와 매력 포인트, 키워드가 핵심 도메인이다.
+ */
 @Service
 @RequiredArgsConstructor
 public class LibraryEvaluationApplication {
@@ -29,6 +41,12 @@ public class LibraryEvaluationApplication {
     private final AttractivePointService attractivePointService;
     private final KeywordService keywordService;
 
+    /**
+     * 서재 평가를 생성한다.
+     *
+     * @param user    사용자 객체
+     * @param request UserNovelCreateRequest
+     */
     @Transactional
     public void createEvaluation(User user, UserNovelCreateRequest request) {
         Novel novel = novelService.getNovelOrException(request.novelId());
@@ -44,6 +62,13 @@ public class LibraryEvaluationApplication {
         }
     }
 
+    /**
+     * 서재 평가를 불러온다.
+     *
+     * @param user    사용자 객체
+     * @param novelId 소설 ID
+     * @return UserNovelGetResponse
+     */
     public UserNovelGetResponse getEvaluation(User user, Long novelId) {
         Novel novel = novelService.getNovelOrException(novelId);
         UserNovel userNovel = libraryService.getUserNovelOrNull(user, novel);
@@ -58,6 +83,24 @@ public class LibraryEvaluationApplication {
         return UserNovelGetResponse.of(novel, userNovel, attractivePoints, keywords);
     }
 
+    /**
+     * 서재 평가를 업데이트한다.
+     *
+     * @param user    사용자 객체
+     * @param novelId 소설 ID
+     * @param request UserNovelUpdateRequest
+     */
+    @Transactional
+    public void updateEvaluation(User user, Long novelId, UserNovelUpdateRequest request) {
+        UserNovel userNovel = libraryService.getUserNovelOrException(user, novelId);
+
+        userNovel.updateUserNovel(request.userNovelRating(), request.status(), request.startDate(), request.endDate());
+
+        updateAttractivePoints(userNovel, request.attractivePoints());
+
+        updateKeywords(userNovel, request.keywordIds());
+    }
+
     // TODO: 리팩토링 대상 Fetch Lazy 수정
     private List<String> getStringAttractivePoints(UserNovel userNovel) {
         return userNovel.getUserNovelAttractivePoints().stream()
@@ -70,6 +113,82 @@ public class LibraryEvaluationApplication {
         return userNovel.getUserNovelKeywords().stream()
                 .map(userNovelKeyword -> KeywordGetResponse.of(userNovelKeyword.getKeyword()))
                 .toList();
+    }
+
+    private void updateAttractivePoints(UserNovel userNovel, List<String> attractivePoints) {
+        Map<AttractivePoint, UserNovelAttractivePoint> currentPointMap = userNovel.getUserNovelAttractivePoints()
+                .stream()
+                .collect(Collectors.toMap(UserNovelAttractivePoint::getAttractivePoint, it -> it));
+
+        Set<AttractivePoint> requestedPoints = attractivePoints.stream()
+                .map(attractivePointService::getAttractivePointByString)
+                .collect(Collectors.toSet());
+
+        addUserNovelAttractivePoints(userNovel, currentPointMap, requestedPoints);
+        deleteUserNovelAttractivePoints(userNovel, currentPointMap, requestedPoints);
+    }
+
+    private void updateKeywords(UserNovel userNovel, List<Integer> keywordIds) {
+        Map<Keyword, UserNovelKeyword> currentKeywordMap = userNovel.getUserNovelKeywords()
+                .stream()
+                .collect(Collectors.toMap(UserNovelKeyword::getKeyword, it -> it));
+
+        Set<Keyword> requestedKeywords = keywordIds.stream()
+                .map(keywordService::getKeywordOrException)
+                .collect(Collectors.toSet());
+
+        addUserNovelKeywords(userNovel, currentKeywordMap, requestedKeywords);
+        deleteUserNovelKeywords(userNovel, currentKeywordMap, requestedKeywords);
+    }
+
+    private void addUserNovelKeywords(UserNovel userNovel,
+                                      Map<Keyword, UserNovelKeyword> currentKeywordMap,
+                                      Set<Keyword> requestedKeywords) {
+        for (Keyword requested : requestedKeywords) {
+            if (!currentKeywordMap.containsKey(requested)) {
+                keywordService.createNovelKeyword(userNovel, requested);
+            }
+        }
+    }
+
+    private void deleteUserNovelKeywords(UserNovel userNovel,
+                                         Map<Keyword, UserNovelKeyword> currentKeywordMap,
+                                         Set<Keyword> requestedKeywords) {
+        List<UserNovelKeyword> toDelete = new ArrayList<>();
+        for (Map.Entry<Keyword, UserNovelKeyword> entry : currentKeywordMap.entrySet()) {
+            if (!requestedKeywords.contains(entry.getKey())) {
+                toDelete.add(entry.getValue());
+            }
+        }
+        if (!toDelete.isEmpty()) {
+            userNovel.getUserNovelKeywords().removeAll(toDelete);
+            userNovel.touch();
+        }
+    }
+
+    private void addUserNovelAttractivePoints(UserNovel userNovel,
+                                              Map<AttractivePoint, UserNovelAttractivePoint> currentPointMap,
+                                              Set<AttractivePoint> requestedPoints) {
+        for (AttractivePoint requested : requestedPoints) {
+            if (!currentPointMap.containsKey(requested)) {
+                attractivePointService.createUserNovelAttractivePoint(userNovel, requested);
+            }
+        }
+    }
+
+    private void deleteUserNovelAttractivePoints(UserNovel userNovel,
+                                                 Map<AttractivePoint, UserNovelAttractivePoint> currentPointMap,
+                                                 Set<AttractivePoint> requestedPoints) {
+        List<UserNovelAttractivePoint> toDelete = new ArrayList<>();
+        for (Map.Entry<AttractivePoint, UserNovelAttractivePoint> entry : currentPointMap.entrySet()) {
+            if (!requestedPoints.contains(entry.getKey())) {
+                toDelete.add(entry.getValue());
+            }
+        }
+        if (!toDelete.isEmpty()) {
+            userNovel.getUserNovelAttractivePoints().removeAll(toDelete);
+            userNovel.touch();
+        }
     }
 
 }
