+++ src/main/java/org/websoso/WSSServer/library/service/AttractivePointService.java
@@ -33,6 +33,10 @@ public class AttractivePointService {
         return getAttractivePointOrException(lowerCaseRequest);
     }
 
+    public void createUserNovelAttractivePoint(UserNovel userNovel, AttractivePoint attractivePoint) {
+        userNovelAttractivePointRepository.save(UserNovelAttractivePoint.create(userNovel, attractivePoint));
+    }
+
     public void createUserNovelAttractivePoints(UserNovel userNovel, List<String> request) {
         for (String stringAttractivePoint : request) {
             AttractivePoint attractivePoint = getAttractivePointByString(stringAttractivePoint);
