+++ src/main/java/org/websoso/WSSServer/library/service/KeywordService.java
@@ -49,6 +49,10 @@ public class KeywordService {
         return KeywordByCategoryGetResponse.of(categories);
     }
 
+    public void createNovelKeyword(UserNovel userNovel, Keyword keyword) {
+        userNovelKeywordRepository.save(UserNovelKeyword.create(userNovel, keyword));
+    }
+
     public void createNovelKeywords(UserNovel userNovel, List<Integer> request) {
         for (Integer keywordId : request) {
             Keyword keyword = getKeywordOrException(keywordId);
